requirements.txt
python-jose[cryptography]==3.5.0
supabase==2.28.2
scapy==2.7.0
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.128.0
uvicorn==0.40.0
slowapi==0.1.9
